refactor: migrate to appDir
