+++ components/animations/fade.jsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState, useRef, useEffect } from 'react';
 
 import { gsap } from 'gsap';
