--- components/common/head/index.jsx
@@ -1,1 +1,0 @@
-export { default as Head } from './head';
