+++ components/common/index.jsx
@@ -1,6 +1,5 @@
 export * from './circular-progress';
 export * from './footer';
-export * from './head';
 export * from './social-links';
 export * from './title';
 export * from './container';
