+++ lib/hooks/index.js
@@ -1,3 +1,3 @@
-export { default as useAnimation } from './useAnimation';
+export { useAnimation } from './useAnimation';
 export { default as useScroll } from './useScroll';
 export { default as useIntersection } from './useIntersection';
