+++ next.config.js
@@ -47,6 +47,9 @@ const securityHeaders = [
 
 module.exports = {
   reactStrictMode: true,
+  experimental: {
+    appDir: true
+  },
   images: {
     formats: ['image/avif', 'image/webp'],
     domains: ['i.scdn.co', 'avatars.githubusercontent.com'],
