+++ package.json
@@ -19,6 +19,7 @@
     "@mdx-js/loader": "^2.3.0",
     "@octokit/graphql": "^5.0.5",
     "@tailwindcss/typography": "^0.5.9",
+    "@types/node": "18.15.11",
     "@vercel/analytics": "^0.1.11",
     "autoprefixer": "^10.4.14",
     "classnames": "^2.3.2",
@@ -48,6 +49,7 @@
     "tailwindcss": "^3.3.1",
     "terser": "^5.16.9",
     "three": "^0.151.3",
+    "typescript": "5.0.4",
     "url-join": "^5.0.0"
   },
   "author": "mateonunez"
