--- pages/500.jsx
@@ -1,11 +1,0 @@
-import s from 'styles/pages/404.module.css';
-
-export default function Eror500() {
-  return (
-    <>
-      <div className={s.root}>
-        <h1 className="title">Error unexpected. Can you fix me?</h1>
-      </div>
-    </>
-  );
-}
