--- pages/_document.jsx
@@ -1,16 +1,0 @@
-import NextDocument, { Head, Html, Main, NextScript } from 'next/document';
-
-export default class Document extends NextDocument {
-  render() {
-    return (
-      <Html lang="en">
-        <Head />
-        <body className="antialiased">
-          <Main />
-          <NextScript />
-          {/* Insert Google Tag Manager */}
-        </body>
-      </Html>
-    );
-  }
-}
